+++ js/eclipse-text-image-with-fx.js
@@ -14,7 +14,7 @@
  * Copyright (c) 2026 r-vage. MIT License.
  */
 import { app } from './comfy/index.js';
-import { createVueColorInputPatcher } from './eclipse-color-picker-utils.js';
+import { createVueColorInputPatcher, installClassicColorDraw } from './eclipse-color-picker-utils.js';
 import { canvasDirtyBatcher, createWidgetVisibilityManager, isConfiguringGraph, isVueMode, notifyVue, smartResize } from './eclipse-widget-performance-utils.js';
 
 const NODE_NAME = 'Text Image with FX [Eclipse]';
@@ -138,7 +138,7 @@ app.registerExtension({
                 };
 
                 // Custom draw: styled row with swatch
-                w.draw = function (ctx, _node, widgetWidth, y, H) {
+                installClassicColorDraw(w, function (ctx, _node, widgetWidth, y, H) {
                     ctx.save();
                     const hex = w.value || '#000000';
                     const margin = 15;
@@ -171,7 +171,7 @@ app.registerExtension({
                     ctx.lineWidth = 1;
                     ctx.stroke();
                     ctx.restore();
-                };
+                });
             }
 
             // Workflow restore